Move constant to constants & change to tsx to ts
