+++ src/constants/error-codes.ts
@@ -1,0 +1,8 @@
+export const ErrorCodes = {
+  INVALID_EMAIL: "auth/invalid-email",
+  USER_DELETED: "auth/user-not-found",
+  INVALID_PASSWORD: "auth/wrong-password",
+  EMPTY_PASSWORD: "auth/missing-password",
+  BAD_PASSWORD: "auth/weak-password",
+  REQUIRE_RECENT_LOGIN: "auth/requires-recent-login",
+}
