+++ src/helpers/sign-in.helper.ts
@@ -1,11 +1,4 @@
-export const ErrorCodes = {
-  INVALID_EMAIL: "auth/invalid-email",
-  USER_DELETED: "auth/user-not-found",
-  INVALID_PASSWORD: "auth/wrong-password",
-  EMPTY_PASSWORD: "auth/missing-password",
-  BAD_PASSWORD: "auth/weak-password",
-  REQUIRE_RECENT_LOGIN: "auth/requires-recent-login",
-}
+import { ErrorCodes } from "@/constants/error-codes"
 
 export const mapAuthErrorCodeToErrorMessage = (
   errorCode: string | null
